+++ bigtable/unit_tests/bigtable/test_cluster.py
@@ -187,7 +187,7 @@ class TestCluster(unittest.TestCase):
         self.assertNotEqual(cluster1, cluster2)
 
     def test_reload(self):
-        from unit_tests._testing import _FakeStub
+        from unit_tests.bigtable._testing import _FakeStub
         from google.cloud.bigtable.cluster import DEFAULT_SERVE_NODES
 
         SERVE_NODES = 31
@@ -232,7 +232,7 @@ class TestCluster(unittest.TestCase):
         from google.cloud.operation import Operation
         from google.cloud.bigtable._generated import (
             bigtable_instance_admin_pb2 as messages_v2_pb2)
-        from unit_tests._testing import _FakeStub
+        from unit_tests.bigtable._testing import _FakeStub
 
         SERVE_NODES = 4
         client = _Client(self.PROJECT)
@@ -281,7 +281,7 @@ class TestCluster(unittest.TestCase):
             instance_pb2 as data_v2_pb2)
         from google.cloud.bigtable._generated import (
             bigtable_instance_admin_pb2 as messages_v2_pb2)
-        from unit_tests._testing import _FakeStub
+        from unit_tests.bigtable._testing import _FakeStub
         from google.cloud.bigtable.cluster import _UPDATE_CLUSTER_METADATA_URL
 
         NOW = datetime.datetime.utcnow()
@@ -339,7 +339,7 @@ class TestCluster(unittest.TestCase):
 
     def test_delete(self):
         from google.protobuf import empty_pb2
-        from unit_tests._testing import _FakeStub
+        from unit_tests.bigtable._testing import _FakeStub
 
         client = _Client(self.PROJECT)
         instance = _Instance(self.INSTANCE_ID, client)
